Refactor cryptarchia state into its own module.
