+++ nomos-services/cryptarchia-consensus/src/lib.rs
@@ -3,9 +3,11 @@ mod leadership;
 mod messages;
 pub mod network;
 mod relays;
+mod states;
 pub mod storage;
+
 use core::fmt::Debug;
-use std::{collections::BTreeSet, hash::Hash, marker::PhantomData, path::PathBuf};
+use std::{collections::BTreeSet, hash::Hash, path::PathBuf};
 
 use cryptarchia_engine::Slot;
 use futures::StreamExt;
@@ -32,7 +34,6 @@ use nomos_time::{SlotTick, TimeService, TimeServiceMessage};
 use overwatch::{
     services::{
         relay::{OutboundRelay, Relay, RelayMessage},
-        state::ServiceState,
         ServiceCore, ServiceData, ServiceId,
     },
     DynError, OpaqueServiceStateHandle,
@@ -51,6 +52,10 @@ use tracing_futures::Instrument;
 use crate::{
     leadership::Leader,
     relays::CryptarchiaConsensusRelays,
+    states::{
+        CryptarchiaConsensusState, CryptarchiaInitialisationStrategy, GenesisRecoveryStrategy,
+        SecurityRecoveryStrategy,
+    },
     storage::{adapters::StorageAdapter, StorageAdapter as _},
 };
 
@@ -542,12 +547,14 @@ where
 
         // TODO: Move to function
         let mut cryptarchia = match self.initial_state.recovery_strategy() {
-            CryptarchiaRecoveryStrategy::NoRecovery => {
-                info!("Building Cryptarchia from genesis without recovery...");
+            CryptarchiaInitialisationStrategy::Genesis => {
+                info!("Building Cryptarchia from genesis...");
                 Cryptarchia::new(genesis_id, genesis_state, config)
             }
-            CryptarchiaRecoveryStrategy::Genesis(GenesisRecoveryStrategy { tip }) => {
-                info!("Building Cryptarchia from genesis with recovery...");
+            CryptarchiaInitialisationStrategy::RecoveryFromGenesis(GenesisRecoveryStrategy {
+                tip,
+            }) => {
+                info!("Recovering Cryptarchia from genesis...");
                 Self::build_cryptarchia_from_recovery(
                     tip,
                     genesis_id,
@@ -559,14 +566,14 @@ where
                 )
                 .await
             }
-            CryptarchiaRecoveryStrategy::Security(strategy) => {
+            CryptarchiaInitialisationStrategy::RecoveryFromSecurity(strategy) => {
                 let SecurityRecoveryStrategy {
                     tip,
                     security_block_id,
                     security_ledger_state,
                 } = *strategy;
 
-                info!("Building Cryptarchia from security block with recovery...");
+                info!("Recovering Cryptarchia from security...");
                 info!("Leader is out of date, updating from genesis until security block.");
                 // TODO: OPTIMIZE: Fetch only headers
                 // TODO: Move to function
@@ -685,129 +692,6 @@ where
             // Probably related to too many generics.
         }.instrument(span!(Level::TRACE, CRYPTARCHIA_ID)).await;
         Ok(())
-    }
-}
-
-pub struct GenesisRecoveryStrategy {
-    tip: HeaderId,
-}
-
-pub struct SecurityRecoveryStrategy {
-    tip: HeaderId,
-    security_block_id: HeaderId,
-    security_ledger_state: LedgerState,
-}
-
-pub enum CryptarchiaRecoveryStrategy {
-    NoRecovery,
-    Genesis(GenesisRecoveryStrategy),
-    Security(Box<SecurityRecoveryStrategy>),
-}
-
-#[derive(Clone, Serialize, Deserialize)]
-pub struct CryptarchiaConsensusState<
-    TxS,
-    BxS,
-    NetworkAdapterSettings,
-    BlendAdapterSettings,
-    TimeBackendSettings,
-> {
-    tip: Option<HeaderId>,
-    security_block: Option<HeaderId>,
-    security_ledger_state: Option<LedgerState>,
-    _txs: PhantomData<TxS>,
-    _bxs: PhantomData<BxS>,
-    _network_adapter_settings: PhantomData<NetworkAdapterSettings>,
-    _blend_adapter_settings: PhantomData<BlendAdapterSettings>,
-    _time_backend_settings: PhantomData<TimeBackendSettings>,
-}
-
-impl<TxS, BxS, NetworkAdapterSettings, BlendAdapterSettings, TimeBackendSettings>
-    CryptarchiaConsensusState<
-        TxS,
-        BxS,
-        NetworkAdapterSettings,
-        BlendAdapterSettings,
-        TimeBackendSettings,
-    >
-{
-    #[must_use]
-    pub const fn new(
-        tip: Option<HeaderId>,
-        security_block: Option<HeaderId>,
-        security_ledger_state: Option<LedgerState>,
-    ) -> Self {
-        Self {
-            tip,
-            security_block,
-            security_ledger_state,
-            _txs: PhantomData,
-            _bxs: PhantomData,
-            _network_adapter_settings: PhantomData,
-            _blend_adapter_settings: PhantomData,
-            _time_backend_settings: PhantomData,
-        }
-    }
-
-    pub(crate) fn from_cryptarchia(cryptarchia: &Cryptarchia) -> Self {
-        let security_block_header = cryptarchia.consensus.get_security_block_header_id();
-        let security_ledger_state = security_block_header
-            .and_then(|header| cryptarchia.ledger.state(&header))
-            .cloned();
-
-        Self::new(
-            Some(cryptarchia.tip()),
-            security_block_header,
-            security_ledger_state,
-        )
-    }
-
-    pub fn can_recover(&self) -> bool {
-        // This only checks whether tip is defined, as that's a state variable that
-        // should always exist. Other attributes might not be present.
-        self.tip.is_some()
-    }
-
-    fn can_recover_from_security(&self) -> bool {
-        self.can_recover() && self.security_block.is_some() && self.security_ledger_state.is_some()
-    }
-
-    pub fn recovery_strategy(&self) -> CryptarchiaRecoveryStrategy {
-        if self.can_recover_from_security() {
-            let strategy = SecurityRecoveryStrategy {
-                tip: self.tip.expect("tip not available"),
-                security_block_id: self.security_block.expect("security block not available"),
-                security_ledger_state: self
-                    .security_ledger_state
-                    .clone()
-                    .expect("security ledger state not available"),
-            };
-            CryptarchiaRecoveryStrategy::Security(Box::new(strategy))
-        } else if self.can_recover() {
-            let strategy = GenesisRecoveryStrategy {
-                tip: self.tip.expect("tip not available"),
-            };
-            CryptarchiaRecoveryStrategy::Genesis(strategy)
-        } else {
-            CryptarchiaRecoveryStrategy::NoRecovery
-        }
-    }
-}
-
-impl<TxS, BxS, NetworkAdapterSettings, BlendAdapterSettings, TimeBackendSettings> ServiceState
-    for CryptarchiaConsensusState<
-        TxS,
-        BxS,
-        NetworkAdapterSettings,
-        BlendAdapterSettings,
-        TimeBackendSettings,
-    >
-{
-    type Settings = CryptarchiaSettings<TxS, BxS, NetworkAdapterSettings, BlendAdapterSettings>;
-    type Error = Error;
-
-    fn from_settings(_settings: &Self::Settings) -> Result<Self, Self::Error> {
-        Ok(Self::new(None, None, None))
     }
 }
